Fixed warnings about arithmetic conversions

+++ OpenGLSandbox/main.cpp
@@ -157,8 +157,8 @@ void mouse_callback(GLFWwindow* window, double xpos, double ypos)
         pitch = -89.0f;
 
     glm::vec3 front;
-    front.x = cos(glm::radians(pitch)) * cos(glm::radians(yaw));
-    front.y = sin(glm::radians(pitch));
-    front.z = cos(glm::radians(pitch)) * sin(glm::radians(yaw));
+    front.x = static_cast<float>(cos(glm::radians(pitch)) * cos(glm::radians(yaw)));
+    front.y = static_cast<float>(sin(glm::radians(pitch)));
+    front.z = static_cast<float>(cos(glm::radians(pitch)) * sin(glm::radians(yaw)));
     cameraFront = glm::normalize(front);
 }
